frontend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Template Engine
Jinja2==3.1.2

# WSGI Server
uWSGI==2.0.23

# Utilities
python-dotenv==1.0.0
requests==2.31.0
